fix(types): avoid float overflow in BBox3.size for empty/degenerate box
An empty Voxels yields an inverted bbox (FLT_MAX/-FLT_MAX); max-min overflowed float32 (-inf + RuntimeWarning). Compute size in float64.

+++ src/picogk/types.py
@@ -62,7 +62,8 @@ class BBox3:
 
     @property
     def size(self) -> np.ndarray:
-        return self.max - self.min
+        # float64 avoids overflow for the degenerate "empty" box (FLT_MAX/-FLT_MAX)
+        return self.max.astype(np.float64) - self.min.astype(np.float64)
 
     @property
     def center(self) -> np.ndarray:
